make basic functionality work for lakefs

+++ catalog/cataloger_create_entry.go
@@ -24,7 +24,7 @@ func (c *cataloger) CreateEntry(ctx context.Context, repository, branch string, 
 	return err
 }
 
-func insertNewEntry(tx db.Tx, branchID int, entry *Entry) error {
+func insertNewEntry(tx db.Tx, branchID int64, entry *Entry) error {
 	_, err := tx.Exec(`INSERT INTO entries (branch_id,path,physical_address,checksum,size,metadata) VALUES ($1,$2,$3,$4,$5,$6)
 			ON CONFLICT (branch_id,path,min_commit)
 			DO UPDATE SET physical_address=$3, checksum=$4, size=$5, metadata=$6, max_commit=$7`,
